requirements.txt
-e .
aws-cdk.aws-ec2
aws-cdk.aws-autoscaling
aws-cdk.aws-elasticloadbalancingv2